moving apis to the proper place
